+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add `size` to glyphs and texts to get effective font size (still not
   entirely accurate when there is rotation or skewing)
 - Support PDF 2.0 `Length` attribute on inline images
+- Add `font` property to documents and pages
 - BREAKING: `find` and `find_all` in structure search by standard
   structure types (roles)
 - BREAKING: `parent_tree` moved to `playa.structure.Tree`